Added/Renamed/Removed java and resource files for fragments.
