+++ app/src/main/java/com/arom/jobzi/AdminActivity.java
@@ -1,6 +1,7 @@
 package com.arom.jobzi;
 
 import android.os.Bundle;
+import android.support.design.widget.TabLayout;
 import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
@@ -10,30 +11,10 @@ import com.arom.jobzi.user.UsersFragment;
 
 public class AdminActivity extends AppCompatActivity {
 
-    /**
-     * The {@link android.support.v4.view.PagerAdapter} that will provide
-     * fragments for each of the sections. We use a
-     * {@link FragmentPagerAdapter} derivative, which will keep every
-     * loaded fragment in memory. If this becomes too memory intensive, it
-     * may be best to switch to a
-     * {@link android.support.v4.app.FragmentStatePagerAdapter}.
-     */
-
-    /**
-     * The {@link ViewPager} that will host the section contents.
-     */
-    private ViewPager mViewPager;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_admin);
-
-        ServicesFragment servicesFragment = new ServicesFragment();
-        getSupportFragmentManager().beginTransaction().add(R.id.container, servicesFragment).commit();
-
-        UsersFragment usersFragment = new UsersFragment();
-        getSupportFragmentManager().beginTransaction().add(R.id.container, usersFragment).commit();
 
     }
 }
