+++ app/src/main/res/layout/fragment_services.xml
@@ -7,22 +7,4 @@
     android:layout_height="match_parent"
     tools:context=".AdminActivity$PlaceholderFragment">
 
-    <TextView
-        android:id="@+id/section_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/activity_horizontal_margin"
-        android:layout_marginTop="@dimen/activity_vertical_margin"
-        android:layout_marginEnd="@dimen/activity_horizontal_margin"
-        android:layout_marginBottom="@dimen/activity_vertical_margin"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/constraintLayout"
-        tools:layout_constraintLeft_creator="1"
-        tools:layout_constraintTop_creator="1" />
-
-    <ListView
-        android:id="@+id/fragmentuServiceList"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
 </android.support.constraint.ConstraintLayout>
