+++ app/src/main/res/layout/fragment_user_list.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragmentUserList"
     android:name="com.arom.jobzi.fragment.UserFragment"
@@ -7,5 +7,12 @@
     android:layout_height="match_parent"
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
-    tools:context=".fragment.UserFragment"
-    tools:listitem="@layout/user_item" />
+    tools:context=".fragment.UserListFragment"
+    tools:listitem="@layout/user_item" >
+
+	<ListView
+		android:id="@+id/userListView"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"></ListView>
+
+</FrameLayout>
