+++ app/src/main/res/values/strings.xml
@@ -1,16 +1,15 @@
 <resources>
     <string name="app_name">JobZi</string>
     <string name="title_activity_welcome">WelcomeActivity</string>
+    <string name="title_activity_signup">SignupActivity</string>
+    <string name="title_activity_admin">AdminActivity</string>
     <string name="email">Email</string>
     <string name="username">Username</string>
     <string name="password">Password</string>
-    <string name="title_activity_signup">SignupActivity</string>
-    <string name="title_activity_admin">AdminActivity</string>
     <string name="user_welcome_banner">Welcome! You are signed in as %s (%s)</string>
     <string name="add_service">Add Service</string>
     <string name="logout">Logout</string>
     <string name="action_settings">Action Settings</string>
+    <string name="services_label">Services</string>
 
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
